# prototype/1. Technical_ EU STR - Code Single Digital Entry Point/str-ap-internal-0.0.49 (update February 2025)/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A gateway for the electronic transmission of data between online short-term rental platforms and competent authorities, ensuring timely, reliable and efficient data sharing processes\nPriority of development is: 1. listings, 2. orders, 3. activity, 4. area\n\u003cbr\u003e\nTo obtain API credentials, please contact: wouter.travers@pwc.com via e-mail",
        "title": "EU STR - Single Digital Entry Point",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Wouter Travers",
            "email": "wouter.travers@pwc.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/ca/data-area": {
            "post": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "write"
                        ]
                    }
                ],
                "description": "Upload a new data area shapefile to the server.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ca"
                ],
                "summary": "Upload a new data area shapefile",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Shapefile to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id of the Competent Authority",
                        "name": "competentAuthorityId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the Competent Authority",
                        "name": "competentAuthorityName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing mandatory parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "422": {
                        "description": "Invalid file format or data",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
        "/ca/data-area/{luid}": {
            "delete": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "write"
                        ]
                    }
                ],
                "description": "Delete a data shapefile from the server based on the LUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ca"
                ],
                "summary": "Delete a data area shapefile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LUID of the shapefile to delete",
                        "name": "luid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/ca/str-area": {
            "post": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "write"
                        ]
                    }
                ],
                "description": "Upload a new STR area shapefile to the server.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ca"
                ],
                "summary": "Upload a new STR area shapefile",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Shapefile to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id of the Competent Authority",
                        "name": "competentAuthorityId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the Competent Authority",
                        "name": "competentAuthorityName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing mandatory parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "422": {
                        "description": "Invalid file format or data",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
        "/ca/str-area/{luid}": {
            "delete": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "write"
                        ]
                    }
                ],
                "description": "Delete the STR shapefile from the server based on the LUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ca"
                ],
                "summary": "Delete a STR area shapefile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LUID of the shapefile to delete",
                        "name": "luid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/str/data-area": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "Retrieve a list of available Data shapefiles.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "str"
                ],
                "summary": "Get a list of available data-shapefiles (shapefiles for which you want to receive activity data for)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.ShapefileMetadata"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
        "/str/data-area/{luid}": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "Retrieve the data shapefile by its ID and download it.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "str"
                ],
                "summary": "Download the shapefile for which you want to receive activity data.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shapefile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Shapefile not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/str/str-area": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "Retrieve a list of available STR shapefiles.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "str"
                ],
                "summary": "Get a list of available shapefiles (shapefiles where the STR regulation is applicable)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.ShapefileMetadata"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.BadRequestError": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Error message for bad requests",
                    "type": "string",
                    "example": "Wrong data format!"
                }
            }
        },
        "handler.DeleteResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Status of the delete operation",
                    "type": "string"
                }
            }
        },
...
        "handler.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message for internal server errors",
                    "type": "string",
                    "example": "An unexpected error occurred"
                }
            }
        },
...
        "handler.NotFoundError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message for resource not found",
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        },
...
        "handler.UnauthorizedError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Error message for unauthorized access",
                    "type": "string",
                    "example": "JWT is invalid"
                }
            }
        },
...
        "handler.UploadResponse": {
            "type": "object",
            "properties": {
                "competentAuthorityId_area": {
                    "type": "string",
                    "example": "competentAuthorityId_area"
                },
                "competentAuthorityName_area": {
                    "type": "string",
                    "example": "competentAuthorityName_area"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://tt-dp-dev.eu.auth0.com/authorize?audience=https://str.eu",
            "tokenUrl": "https://tt-dp-dev.eu.auth0.com/oauth/token",
            "scopes": {
                "str": "Grants read access"
            }
        }
    },
...
}